backend/requirements.txt
Django>=3.0,<4.0
psycopg2>=2.7,<3.0
djangorestframework>=3.0,<4.0
django-cors-headers>=3.2,<4.0
recipe-scrapers>=6.0,<7.0
wget>=3.0,<4.0
Pillow>=7.1,<8.0
imapy>=1.2,<2.0
celery>=4.0,<5.0
django-health-check>=3.14